feat(#584): rename profile to people
