+++ packages/epics/src/people/README.md
