+++ packages/epics/src/people/components/author.tsx
