+++ packages/epics/src/people/components/person-avatar.tsx
