+++ packages/epics/src/people/components/person-head.tsx
